runtime: normalize (*sigctxt).fault() type
(*sigctxt).fault() currently returns either uintptr, uint32, or uint64
depending on the platform. Make them all return uintptr.

For #10958 (but a nice change on its own).

Reviewed-on: https://go-review.googlesource.com/68015
Run-TryBot: Austin Clements <austin@google.com>
TryBot-Result: Gobot Gobot <gobot@golang.org>

+++ src/runtime/signal_nacl_arm.go
@@ -36,7 +36,7 @@ func (c *sigctxt) lr() uint32  { return c.regs().lr }
 func (c *sigctxt) pc() uint32 { return c.regs().pc }
 
 func (c *sigctxt) cpsr() uint32    { return c.regs().cpsr }
-func (c *sigctxt) fault() uint32   { return ^uint32(0) }
+func (c *sigctxt) fault() uintptr  { return ^uintptr(0) }
 func (c *sigctxt) trap() uint32    { return ^uint32(0) }
 func (c *sigctxt) error() uint32   { return ^uint32(0) }
 func (c *sigctxt) oldmask() uint32 { return ^uint32(0) }
